matrice en float

+++ fonction.h
@@ -19,7 +19,7 @@ void binaire(image* im,int seuil);
 void contraste(image* img);
 void contour(image* img);
 void floutage(image* img);
-void convolution(image* img, int** matrice,image* img2);
+void convolution(image* img, float** matrice,image* img2);
 void histogramme(image* im,int* tab);
 void miroir(image* im);
 void negatif(image* im);
